requirements_1.txt
# SKHU Agent V1.0 - Requirements
# Python >= 3.10

# ── Web UI ────────────────────────────────────────────────────────────────────
streamlit>=1.55.0

# ── Notebook Parsing ──────────────────────────────────────────────────────────
nbformat>=5.10.4

# ── LangChain / LangGraph ─────────────────────────────────────────────────────
langchain>=1.2.12
langchain-core>=1.2.18
langchain-openai>=1.1.11
langchain-community>=0.4.1
langchain-classic>=1.0.2
langchain-text-splitters>=1.1.1
langgraph>=1.1.2

# ── Vector Store ──────────────────────────────────────────────────────────────
faiss-cpu>=1.13.2

# ── Keyword Search ────────────────────────────────────────────────────────────
rank-bm25>=0.2.2
kiwipiepy>=0.18.0

# ── Graph ─────────────────────────────────────────────────────────────────────
networkx>=3.6.1

# ── Data / Utils ──────────────────────────────────────────────────────────────
pandas>=2.3.3
numpy>=2.4.3
Pillow>=12.1.1

# ── OpenAI ────────────────────────────────────────────────────────────────────
openai>=2.26.0
tiktoken>=0.12.0

# ── Markdown rendering ───────────────────────────────────────────────────────
Markdown>=3.7

# ── Environment ───────────────────────────────────────────────────────────────
# env.txt 파일을 직접 파싱하므로 python-dotenv 불필요
